Initialize the geist actor type & refactor the HeroData data model
