+++ module/data/Actor/Geist.mjs
@@ -1,0 +1,3 @@
+import { EntityData } from "./Entity.mjs";
+
+export class GeistData extends EntityData {};
